Distributed 3PC on a generic 2PC framework
This impl leverages ZK to orchestrate the distributed operation,
and relies on 2PC operations on both the coordinator and
each of the cohort members along with a local timer to
create a distributed 3PC.

The framework is based on a builder pattern to allow easier
testing and error injection on demand. Each side (coordinator
and cohort) has its own builder that can be set via constructor.

9/15/12
 - rebased ontop of current error-handling on snapshot branch
 - fixed compile errors from rebase onto latest error handling
9/16/12
 - General cleanup and test fixes
9/18/12
 - Removing some of the typing.
 -- in short, all distributed 3PC are going to be distributedcommitExceptions.
9/21/12
 - Adding ability to fail to build a cohort operation